requirements.txt
absl-py==2.1.0
accelerate==0.30.1
addict==2.4.0
aiobotocore==2.13.0
aiofiles==23.2.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==3.7.1
art==6.2
asttokens==2.4.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
-e git+https://github.com/OpenAccess-AI-Collective/axolotl@3f1f5e33120b75f83736d75eb3bea0a6dad5424c#egg=axolotl
bidict==0.23.1
bitsandbytes==0.43.1
botocore==1.34.106
cachetools==5.3.3
certifi==2024.6.2
chainlit==1.1.0
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
cmake==3.29.5.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.5.14
datasets==2.19.1
debugpy==1.8.1
decorator==5.1.1
deepspeed @ git+https://github.com/microsoft/DeepSpeed.git@bc48371c5e1fb8fd70fc79285e66201dbb65679b
deepspeed-kernels==0.0.1.dev1698255861
Deprecated==1.2.14
diffusers==0.29.0
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
email_validator==2.1.2
et-xmlfile==1.1.0
evaluate==0.4.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.3
fastapi-cli==0.0.4
fastapi-socketio==0.0.10
fastcore==1.5.46
ffmpy==0.3.2
filelock==3.15.1
filetype==1.2.0
fire==0.6.0
flash-attn==2.5.8
fonttools==4.53.0
frozenlist==1.4.1
fschat @ git+https://github.com/lm-sys/FastChat.git@27a05b04a35510afb1d767ae7e5990cbd278f8fe
fsspec==2024.3.1
gcsfs==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
gradio==3.50.2
gradio_client==0.6.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hf_transfer==0.1.6
hjson==3.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
langchain-openai==0.1.9
langchain-qdrant==0.1.1
langchain-text-splitters==0.2.1
langsmith==0.1.81
Lazify==0.4.0
lazy_loader==0.4
literalai==0.0.601
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nh3==0.2.17
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.35.3
openpyxl==3.1.4
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
optimum==1.16.2
orjson==3.10.5
packaging==23.2
pandas==2.2.2
parso==0.8.4
peft==0.11.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
portalocker==2.10.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pytorch-ignite==0.5.0.post2
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qdrant-client==1.9.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
responses==0.18.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3fs==2024.3.1
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.2.2
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.5.1
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
svgwrite==1.4.3
sympy==1.12.1
syncer==2.0.3
tenacity==8.4.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.22
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.1
torch==2.3.0
torchinfo==1.8.0
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.1
triton==2.3.0
trl @ git+https://github.com/huggingface/trl.git@f18253bf2d747f68acc9cd89da95c85ebf59dbb9
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tyro==0.8.4
tzdata==2024.1
ujson==5.10.0
unsloth @ git+https://github.com/unslothai/unsloth.git@a2ee56813ed67b7f5336793cbca84442a94140fd
uptrace==1.24.0
urllib3==2.2.1
uvicorn==0.25.0
uvloop==0.19.0
wandb==0.17.2
watchfiles==0.20.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
wsproto==1.2.0
xformers==0.0.26.post1
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
zstandard==0.22.0
